+++ .gitignore
@@ -2,7 +2,9 @@
 predictor_pol/__pycache__/*
 predictor_pol/.idea
 predictor_pol/predictor_prod.db
-predictor_pol/secrets.json
+predictor_pol/json_data/secrets.json
 .DS_Store
 *.joblib
 *.csv
+__pycache__
+.flake8
