+++ predictor_pol/json_data/candidatos.json
